% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agxgman.R
\name{agxgman}
\alias{agxgman}
\title{agxgman}
\usage{
agxgman(d, me, symmetric = TRUE, highlight_p = 0.05, legend = "p-value",
  title = NULL, high = "#02021e", low = "blue",
  highlight_high = "yellow", highlight_low = "#fffcd3", file = "agxgman",
  ext = "gif", hgt = 600, wi = 650)
}
\arguments{
\item{d}{data frame, must contain SNP1, CHR1, POS1, SNP2, CHR2, POS2, pvalue, Frame (factor) columns}

\item{me}{optional main effect data frame containing SNP, CHR, POS, pvalue, Frame (factor) columns, optional Shape}

\item{symmetric}{boolean, should the plot be symmetric}

\item{highlight_p}{threshold to diverge gradient color scale, default 0.05, set to "off" for no threshold}

\item{legend}{title for color legend, default "p-value"}

\item{title}{optional string for plot title}

\item{high}{color for high values}

\item{low}{color for low values}

\item{highlight_high}{if highlight_p given, color for max of highlight range}

\item{highlight_low}{if highlight_p given, color for min of highlight range}

\item{file}{file name of saved image}

\item{ext}{file extension "html" or "gif"}

\item{hgt}{height of plot in pixels}

\item{wi}{width of plot in pixels}
}
\value{
html or gif file
}
\description{
Create animated heatmap plots for SNPxSNP or GxG interaction
Dependencies: ggplot2
}
\examples{
agxgman(d, me, symmetric, highlight_p, legend, title, high, low, highlight_high, highlight_low, file, hgt, wi)
}
