% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeman.R
\name{gxeman}
\alias{gxeman}
\title{gxeman}
\usage{
gxeman(d, snpme, envme, highlight_p = 0.05, legend = "p-value",
  title = NULL, high = "#02021e", low = "blue",
  highlight_high = "yellow", highlight_low = "#fffcd3", file = "gxgman",
  hgt = 7, wi = 7.5, res = 300)
}
\arguments{
\item{d}{data frame, must contain SNP, ENV, CHR, POS, pvalue columns, optional Group column (for grouping ENV)}

\item{snpme}{optional main effect data frame containing SNP, CHR, POS, pvalue columns, optional Shape column}

\item{envme}{optional main effect data frame containing ENV and pvalue columns, optional Group column (for grouping ENV)}

\item{highlight_p}{threshold to diverge gradient color scale, default 0.05, set to "off" for no threshold}

\item{legend}{title for color legend, default "p-value"}

\item{title}{optional string for plot title}

\item{high}{color for high values}

\item{low}{color for low values}

\item{highlight_high}{if highlight_p given, color for max of highlight range}

\item{highlight_low}{if highlight_p given, color for min of highlight range}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image
}
\description{
Create heatmap plots for SNPxE or GxE interaction
Dependencies: ggplot2
}
\examples{
gxeman(d, me, symmetric, highlight_p, legend, title, high, low, highlight_high, highlight_low, file, hgt, wi, res)
}
