% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igxgman.R
\name{igxgman}
\alias{igxgman}
\title{igxgman}
\usage{
igxgman(d, me, symmetric = TRUE, highlight_p = 0.05, legend = "p-value",
  title = NULL, high = "#02021e", low = "#1A0F99",
  highlight_high = "yellow", highlight_low = "#fffcd3", dbSNP,
  moreinfo = FALSE, file = "igxgman", hgt = 7, wi = 7.5, res = 300)
}
\arguments{
\item{d}{data frame, must contain SNP1, CHR1, POS1, SNP2, CHR2, POS2, pvalue columns, optional Info column}

\item{me}{optional main effect data frame containing SNP, CHR, POS, pvalue columns, optional Shape and Info column}

\item{symmetric}{boolean, should the plot be symmetric}

\item{highlight_p}{threshold to diverge gradient color scale, default 0.05, set to "off" for no threshold}

\item{legend}{title for color legend, default "p-value"}

\item{title}{optional string for plot title}

\item{high}{color for high values}

\item{low}{color for low values}

\item{highlight_high}{if highlight_p given, color for max of highlight range}

\item{highlight_low}{if highlight_p given, color for min of highlight range}

\item{moreinfo}{includes more information on hover, refers to Info column}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{db}{database to connect to (GWAScatalog or dbSNP), only used when 'me' is provided}
}
\value{
html file
}
\description{
Create interactive heatmap plots for SNPxSNP or GxG interaction
Dependencies: ggplot2
}
\examples{
igxgman(d, me, symmetric, highlight_p, legend, title, high, low, highlight_high, highlight_low, file, hgt, wi)
}
