% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlight.R
\name{nightlight_apply}
\alias{nightlight_apply}
\title{Apply a function to each geometric object}
\usage{
nightlight_apply(nightlight_data, geom, func, ...)
}
\arguments{
\item{nightlight_data}{source path where nightlight data was downloaded}

\item{geom}{geometric object}

\item{func}{function to apply}

\item{...}{arguments passed to \code{func}}
}
\description{
\code{nightlight_apply} applies a function over each geometric object in \code{geom} for each year of nightlight data in \code{nightlight_data}
}
\examples{
library(nightlight)

set the root folder for where you want to download the dataset
NOAA_DATASET_ROOT <- "~/datasets/noaa"

nightlight_download(NOAA_DATASET_ROOT)

nightlight_data <- nightlight_load(NOAA_DATASET_ROOT)

create a polygon for testing
coordinates <- matrix(
  c(
    113.4789, 22.19556,
    113.4811, 22.21748,
    113.4941, 22.24155,
    113.5271, 22.24595,
    113.5481, 22.22261,
    113.5455, 22.22148,
    113.4988, 22.20166,
    113.4842, 22.19775,
    113.4789, 22.19556
  ),
  ncol = 2,
  byrow = TRUE
)
polygon <- Polygon(coordinates)

spatial_polygons <- SpatialPolygons(list(Polygons(list(polygon),1)))

results <- nightlight_apply(nightlight_data, c(spatial_polygons), mean, na.rm = TRUE)

print(results)
}

