% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlight.R
\name{nightlight_download}
\alias{nightlight_download}
\title{Download nighlight data from NOAA}
\usage{
nightlight_download(dest = ".",
  src = "ftp://ftp.ngdc.noaa.gov/STP/DMSP/web_data/v4composites",
  max_retry = 5)
}
\arguments{
\item{dest}{destination folder for the downloaded nightight data}

\item{src}{source URL of NOAA nightlight data}

\item{max_retry}{maximum number of retries}
}
\description{
\code{nightlight_download} downloads nightlight data from NOAA's website and places it the specified destination directory
}
\examples{
library(nightlight)

nightlight_download("~/datasets/noaa")
}

