% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlight.R
\name{nightlight_load}
\alias{nightlight_load}
\title{Load the nightlight dataset and return a vector of raster objects}
\usage{
nightlight_load(src)
}
\arguments{
\item{src}{source folder where nightlight data was downloaded}
}
\description{
\code{nightlight_load} loads nightlight data from the specified destination directory and returns a vector of raster objects
}
\examples{
library(nightlight)

nightlight_load("~/datasets/noaa")
}

